Always use JSON objects as responses
Even if we currently only return a list of identifiers always wrapping
it in an object gives better future extensibility and makes
successful and error responses be of the same kind.

Add simple `IdList` wrapper with a generically named `items` field to
api and use it instead of `String[]` return values.

Adjust controller test to the changed return type.