Implement returning KPI metadata
Add to the API package `MetricMetadata` with metric's name,
description, unit and type.
Add to `MetricDelivery` methods returning metadata for a single
`(groupId, metricId)` and for all metrics in a group
and implement them in the controller using a
`MetricsMetadataProvider`.

Create the new `MetricsMetadataProvider` interface and a dummy
implementation of it.