Remove logback configuration file (closes #21)
The logback configuration file does not contain at the moment any
important changes that Spring Boot cannot handle with application
properties (see
<https://docs.spring.io/spring-boot/docs/current/reference/htmlsingle/#features.logging>).

The properties are more convenient to use for what they can handle,
eg. because they can be set individually for deployment needs and do
not require handling irrelevant parts of the configuration, so
let's get rid of `logback-spring.xml` until it is proven to be needed.