Consistently use resource/metric terminology
Closes #20.
Following the HTTP API changes started in 73aa267 replace internal
uses of 'group' with 'resource' as well.
At the same time back off from the 'metric' -> 'kpi' change started in
that commit and use 'metric' everywhere. To be fully consistent
otherwise would require changes to many class names and to some HTTP
API paths and not every value reported by the service is going to be
'key' anyway.