Add handling of queries for range vectors to the Prometheus client
The interface assumes one result type for one query type.
Actually an instant query (`api/v1/query`) may returns a vector result
or a matrix result depending on the query - queries for a range vector
return a matrix result.

So add a separate `queryMatrix` method to be used when a matrix
response is expected.

Requires giving wider access to `MatrixData` and `MatrixResult` to be
able to actually access the results in other code, the same as was
done for vectors before.