Prepare for handling metrics incompatible with PromQL
Unfortunately we need to handle some metrics which are not exposed in
a way useful for direct querying with PromQL.
For example we may get noncumulative values for each month - so
not a Prometheus counter - which we must aggregate over a period of
time.

For this some case-specific code will be needed on the Java side.
To try to combine this with the json configuration with as little
hardcoding of metric and timeseries names as possible in the Java code
we use the Spring Expression Language.

In json treat a `query` starting with '#' as a SpEL expression,
not a PromQL query, eg. "#calc('some_series', other_param)".

Create `ExpressionContext` class to contain code callable from the
expression and possibly other context necessary during its evaluation.
Give the context access to the Prometheus API.

In `PrometheusMetricsProvider`, when encountering a SpEL query run it,
providing it an instance of the context and expecting a float result
to return as the metric value.