A method of summing of a noncumulative metric over time
Add a `sumPeriods` method to the `ExpressionContext`.
It takes two timeseries and an initial timestamp.
For the range from the initial timestamp to now it sums values of the
first series corresponding to distinct values of the second series.
So for example if the second series is a timestamp increased each time
(and only when) a new value of the first series is provided
this results in the sum of all the provided values in the range.