Handle queries for counter increases from a specific timestamp
We need to be able to calculate `increase` for a range that is not of
a predetermined length but instead starts at a specific point in time.

PromQL has only syntax for the former. So to transform the second case
into the first allow placing a "#R#" token in queries configuration
instead of a range specification.

Before the query is sent to Prometheus the token is replaced by an
actual range in seconds, calculated from the query time and a configured
(in application properties) range start time.

For now a global range start value should be enough. Many of the
affected metrics are updated only rarely (monthly), so while an
increase from the project start time is useful others may not be.