Add a separate Jenkinsfile for the new dockerized Jenkins instance
It should do the same as the original Jenkinsfile plus publish
master snapshots to the artifactory.

Add a `pollSCM` trigger because there is no way to setup the github
webhook yet.

Instead of using the Sonar Jenkins plugin keep the Sonar token as
Jenkins credentials.

Use a Maven settings file `deploy-settings.xml` to pass the
Artifactory credentials from Jenkins to Maven.

Replace the unavailable `warnings` step by the `recordIssues` step
from the warnings-ng plugin.

Use separate stages for things to do for all branches (build, test)
and things to do only on master (Sonar, Artifactory).
Unfortunately this means rerunning the lifecycle twice for master.
Recompilation recognizes that there is nothing to do and we tell maven
to skip tests, but if something else that cannot be skipped and takes
a long time comes up, running maven once in a single stage instead may
be necessary.