Push metrics to the EOSC accounting service
Add EOSC metric definition identifers to metric metadata.
Add EOSC installation identifier to resource metadata
- crate class ResourceMetadata to hold it and the metrics metadata,
before we just used a map for the metrics metadata.

Update the metadata provider reading the metadata mappings JSON to
handle the new metadata.

Enable the use of Scheduled annotations in the application.

Create package eu.openaire.mas.eosc and add components from this
package to the application if the Spring profile "eosc" is active.

Create AccountingApiAuthenticationService to obtain the access token
for the accounting service based on the refresh token.
It calls into EOSC AAI at a specified rate (every 1800s by default -
access tokens are currently valid 3600s) to obtain a new token.

Create AccountingApiMetricsPushService to actually push the metrics to
the EOSC accounting service.
It uses the API client code autogenerated based on the open-api
specification and at a specified rate (every 24h by default) pushes
the current values for all the metrics with specified EOSC
installation id and metric definition id to the accounting service.

Add dependency on jackson-datatype-jsr310, needed so that date
serialization in requests uses the correct format.

Create AccountingApiControlController to be able to perform
operations such as metrics push or token refresh on demand and
not only on schedule.

Add application-eosc.properties file with default values for
non-secret properties required in the eosc profile.

Update README with new metadata mappings format and information about
accounting system integration configuration.